refact(Port) factor out requests and services to inherited Ports
Mostly to clean up Port, have less stuff in the file
